Split TextLayer mutation tests into focused files
